+++ src/idc/filter/imgcls/__init__.py
@@ -1,1 +1,3 @@
 from ._label_from_name import LabelFromName
+from ._label_present import LabelPresent
+