+++ src/idc/filter/imgseg/_label_present.py
@@ -1,0 +1,116 @@
+import argparse
+from typing import List
+
+from seppl.io import Filter
+from wai.logging import LOGGING_WARNING
+
+from idc.api import flatten_list, make_list, ImageSegmentationData
+
+
+class LabelPresent(Filter):
+    """
+    Only forwards images that have the specified label(s) present.
+    """
+
+    def __init__(self, labels: List[str] = None,
+                 logger_name: str = None, logging_level: str = LOGGING_WARNING):
+        """
+        Initializes the filter.
+
+        :param labels: the labels to use
+        :type labels: list
+        :param logger_name: the name to use for the logger
+        :type logger_name: str
+        :param logging_level: the logging level to use
+        :type logging_level: str
+        """
+        super().__init__(logger_name=logger_name, logging_level=logging_level)
+        self.labels = labels
+
+    def name(self) -> str:
+        """
+        Returns the name of the handler, used as sub-command.
+
+        :return: the name
+        :rtype: str
+        """
+        return "label-present-is"
+
+    def description(self) -> str:
+        """
+        Returns a description of the handler.
+
+        :return: the description
+        :rtype: str
+        """
+        return "Only forwards images that have the specified label(s) present."
+
+    def accepts(self) -> List:
+        """
+        Returns the list of classes that are accepted.
+
+        :return: the list of classes
+        :rtype: list
+        """
+        return [ImageSegmentationData]
+
+    def generates(self) -> List:
+        """
+        Returns the list of classes that get produced.
+
+        :return: the list of classes
+        :rtype: list
+        """
+        return [ImageSegmentationData]
+
+    def _create_argparser(self) -> argparse.ArgumentParser:
+        """
+        Creates an argument parser. Derived classes need to fill in the options.
+
+        :return: the parser
+        :rtype: argparse.ArgumentParser
+        """
+        parser = super()._create_argparser()
+        parser.add_argument("--labels", type=str, default=None, help="The labels to use", required=False, nargs="*")
+        return parser
+
+    def _apply_args(self, ns: argparse.Namespace):
+        """
+        Initializes the object with the arguments of the parsed namespace.
+
+        :param ns: the parsed arguments
+        :type ns: argparse.Namespace
+        """
+        super()._apply_args(ns)
+        self.labels = ns.labels
+
+    def initialize(self):
+        """
+        Initializes the processing, e.g., for opening files or databases.
+        """
+        super().initialize()
+        if (self.labels is None) or (len(self.labels) == 0):
+            raise Exception("No label(s) specified!")
+
+    def _do_process(self, data):
+        """
+        Processes the data record(s).
+
+        :param data: the record(s) to process
+        :return: the potentially updated record(s)
+        """
+        result = []
+
+        for item in make_list(data):
+            present = False
+            for label in self.labels:
+                if item.has_layer(label):
+                    present = True
+                    break
+
+            if present:
+                result.append(item)
+            else:
+                self.logger().info("Label(s) '%s' not present, discarding: %s" % (str(self.labels), item.image_name))
+
+        return flatten_list(result)