+++ CHANGES.rst
@@ -23,6 +23,9 @@ Changelog
 - added depth information readers `from-grayscale-dp`, `from-numpy-dp`, `from-csv-dp` and `from-pfm-dp`
 - added depth information writers `to-grayscale-dp`, `to-numpy-dp`, `to-csv-dp` and `to-pfm-dp`
 - added `apply-ext-mask` filter to applying external PNG masks to image containers (image and/or annotations)
+- added `apply-label-mask` filter for applying image segmentation label masks to their base images
+- added `label-present-ic` and `label-present-is` that ensure that certain label(s) are present or otherwise discard the image
+- filter `label-present` was renamed to `label-present-od` but keeping `label-present` as alias for the time being
 
 
 0.0.11 (2025-04-03)